refactor:formatUtil함수 위치 이동 및 직관성 향상을 위한 변수명 변경
